Rename class 'org.hibernate.tool.internal.reveng.TableProcessor' to 'org.hibernate.tool.internal.reveng.reader.TableCollector'
